pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-supervisor>=0.0.30",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
]
